+++ src/main/java/com/kalsym/product/service/controller/StoreProductVariantController.java
@@ -183,6 +183,48 @@ public class StoreProductVariantController {
         response.setStatus(HttpStatus.OK);
         return ResponseEntity.status(response.getStatus()).body(response);
     }
+    
+    @DeleteMapping(path = {"/bulk"}, name = "store-product-variants-delete-by-id", produces = "application/json")
+    @PreAuthorize("hasAnyAuthority('store-product-variants-delete-by-id', 'all') and @customOwnerVerifier.VerifyStore(#storeId)")
+    public ResponseEntity<HttpResponse> deleteStoreProductVariantsByBulk(HttpServletRequest request,
+            @PathVariable String storeId,
+            @PathVariable String productId,
+            @RequestBody ProductVariant[] bodyProductVariantList) {
+        String logprefix = request.getRequestURI();
+        HttpResponse response = new HttpResponse(request.getRequestURI());
+
+        Logger.application.info(ProductServiceApplication.VERSION, logprefix, "storeId: " + storeId);
+
+        Optional<Store> optStore = storeRepository.findById(storeId);
+
+        if (!optStore.isPresent()) {
+            Logger.application.info(Logger.pattern, ProductServiceApplication.VERSION, logprefix, " NOT_FOUND storeId: " + storeId);
+            response.setStatus(HttpStatus.NOT_FOUND);
+            response.setError("store not found");
+            return ResponseEntity.status(response.getStatus()).body(response);
+        }
+        Logger.application.info(Logger.pattern, ProductServiceApplication.VERSION, logprefix, " FOUND storeId: " + storeId);
+
+        Optional<Product> optProdcut = productRepository.findById(productId);
+
+        if (!optProdcut.isPresent()) {
+            Logger.application.info(Logger.pattern, ProductServiceApplication.VERSION, logprefix, "product NOT_FOUND productId: " + productId);
+            response.setStatus(HttpStatus.NOT_FOUND);
+            response.setError("product not found");
+            return ResponseEntity.status(response.getStatus()).body(response);
+        }
+        
+        for (int i=0;i<bodyProductVariantList.length;i++) {
+            ProductVariant bodyProductVariant = bodyProductVariantList[i];
+            Optional<ProductVariant> optProductVariant = productVariantRepository.findById(bodyProductVariant.getId());
+            if (optProductVariant.isPresent()) {
+                productVariantRepository.delete(optProductVariant.get());                
+            }
+        }
+        
+        response.setStatus(HttpStatus.OK);
+        return ResponseEntity.status(response.getStatus()).body(response);
+    }
 
     @PostMapping(path = {""}, name = "store-product-variants-post", produces = "application/json")
     @PreAuthorize("hasAnyAuthority('store-product-variants-post', 'all') and @customOwnerVerifier.VerifyStore(#storeId)")